first time working +-

+++ lib/ui/CustomWidgets/TemplateEditorTreeWidget.py
@@ -1,7 +1,7 @@
 from PyQt6.QtWidgets import  QWidget, QGridLayout, QTreeWidgetItem
 from PyQt6.QtCore import Qt
 from pyodbc import Row
-from lib.xml.transport_template import transport_template_custom_object
+from lib.xml.transport_template import transport_template_custom_object, object_container
 from lib.ui.CustomWidgets.TemplateEditorListWidget import TemplateEditorListWidgetItem
 import copy
 import re
@@ -22,13 +22,38 @@ class TemplateEditorTreeWidgetItem(QTreeWidgetItem):
 
         self.refresh()
 
+
     def deleteObject(self):
-        if self.xml_object.delete_object():
-            return True
-        return False
+        if isinstance(self.xml_object, transport_template_custom_object):
+            if self.xml_object.delete_object():
+                return True
+            else:
+                return False           
+
+        if isinstance(self, TemplateEditorTreeWidgetItem):
+            for i in reversed(range(self.childCount())):
+                child_item = self.child(i)
+                if isinstance(child_item, TE_RelationColumn_TreeWidgetItem):
+                    child_item.delete_relation(child_item.object_data)
+            if self.source_widget is not None:
+                self.source_widget.refresh()
+
+        if isinstance(self, TE_RelationColumn_TreeWidgetItem) and self.object_relations is not None:
+            self.delete_relation(self.object_data)
+            
+        return True
+
+    def delete_relation(self, relation):
+        if relation in self.source_widget.object_relations:
+            self.source_widget.object_relations.remove(relation)
+        return True
+
 
     def refresh(self):
         self.setDisplayName(self.display_name)
+        if isinstance(self.xml_object, object_container):
+            self.xml_object.reset_container_relations()
+            self.application.list_related_objects(self)
 
     @property
     def tooltipText(self):
@@ -110,9 +135,14 @@ class TemplateEditorTreeWidgetItem(QTreeWidgetItem):
     def follow_table(self):
         value = ""
         if isinstance(self.object_data, Row):
-            if self.table_name == "DialogTable":
+            if self.table_name== "DialogTable":
                 value = self.object_data.TableName
         return value
+
+    @property
+    def delete_residual_objects(self):
+        #TODO: add handling for the value based on the UI selection
+        return "1"
 
     @property
     def pk_columns(self):
@@ -138,54 +168,97 @@ class TemplateEditorTreeWidgetItem(QTreeWidgetItem):
                 if not status:
                     follow_up_relations.append(relation)
 
-        print("follow up relations:", follow_up_relations)
-
         for f_relation in follow_up_relations:
             status = self.get_relation_objects(f_relation, selected_objects)
             if not status:
                 print("follow up failed:", f_relation)
-        print(selected_objects)
+
         return selected_objects
 
     def get_relation_objects(self, relation, selected_objects={}):
-        base_object_table = relation["ParentTable"]
-        base_object_value_column = relation["ParentColumn"]
+        BaseTable = relation["TableGroup"]
+        TableRelation = relation["Relation"]
+        ParentTable = relation["ParentTable"]
+        ParentColumn = relation["ParentColumn"]
 
-        table_name = relation["ChildTable"]
-        column_name = relation["ChildColumn"]
+        ChildTable = relation["ChildTable"]
+        ChildColumn = relation["ChildColumn"]
 
-        column_value = self.get_value(base_object_value_column)
-        print(f"Relation: ", relation)
+        recursive = False
+        if TableRelation in [2, 3]:
+            recursive = True
+        column_value = ""
+        values_list = []
+        # print(f"Relation: ", relation)
 
-        if base_object_table != self.table_name and base_object_table in selected_objects.keys():
-            query_values_list = []
-            for related_object in selected_objects[base_object_table]:
-                related_object_columns = self.get_db_object_columns(related_object)
-                if base_object_value_column in related_object_columns:
-                    base_object_value = related_object.__getattribute__(base_object_value_column)
-                    query_values_list.append(str(base_object_value))
-            column_value = "', '".join(query_values_list)
-            print(column_value)
+        if BaseTable == ParentTable and ParentTable == self.table_name:
+            # column_value = self.get_value(ParentColumn)
+            values_list = self.get_db_objects_values(selected_objects, ParentTable, ParentColumn)
+            if len(values_list) > 0:
+                column_value = "', '".join(values_list)
+                self.get_db_objects(ChildTable, ChildColumn, column_value, selected_objects, recursive, ParentColumn)
 
-        query_results = self.get_db_objects(table_name, column_name, column_value)
-        
-        for db_record in query_results:
-            if table_name not in selected_objects.keys():
-                selected_objects[table_name] = [db_record]
-            else:
-                if db_record not in selected_objects[table_name]:
-                    selected_objects[table_name].append(db_record)
+        elif BaseTable == ChildTable and ChildTable == self.table_name:
+            # column_value = self.get_value(ChildColumn)
+            values_list = self.get_db_objects_values(selected_objects, ChildTable, ChildColumn)
+            if len(values_list) > 0:
+                column_value = "', '".join(values_list)
+            self.get_db_objects(ParentTable, ParentColumn, column_value, selected_objects, recursive, ChildColumn)
+        else:
+            """ foreign table relation """
+            values_list = []
+            if BaseTable == ParentTable:
+                values_list = self.get_db_objects_values(selected_objects, ParentTable, ParentColumn)
+                if len(values_list) > 0:
+                    column_value = "', '".join(values_list)
+                    self.get_db_objects(ChildTable, ChildColumn, column_value, selected_objects, recursive, ParentColumn)
+                else:
+                    print("Parent table data not found:", ParentTable)
+
+            if BaseTable == ChildTable:
+                values_list = self.get_db_objects_values(selected_objects, ChildTable, ChildColumn)
+                if len(values_list) > 0:
+                    column_value = "', '".join(values_list)
+                    self.get_db_objects(ParentTable, ParentColumn, column_value, selected_objects, recursive, ChildColumn)
+                else:
+                    print("Parent table data not found:", ParentTable)
         return True
 
-    def get_db_objects(self, table_name, query_column, query_values):
-        selected_objects = []
-        query = f"select * from {table_name} where {query_column} in ('{query_values}')"
+    def get_db_objects_values(self, all_objects_dict, table_name, column_name):
+        values_list = []
+        if table_name in all_objects_dict.keys():
+            related_objects = all_objects_dict[table_name]
+            for record in related_objects:
+                record_columns = self.get_db_object_columns(record)
+                if column_name in record_columns:
+                    column_value = record.__getattribute__(column_name)
+                    if column_value is not None:
+                        values_list.append(column_value)
+        return values_list
 
+
+    def get_db_objects(self, table_name, query_column, query_values, selected_objects_dict={}, recursive=False, recursive_key_column=None):
+
+        query = f"select * from {table_name} where {query_column} in ('{query_values}')"
         query_result = self.db_connection.run_db_query(query)
-        for row in query_result:
-            selected_objects.append(row)
-        print("results:", len(selected_objects), "query:", query)
-        return selected_objects
+
+        # print("results:", len(query_result), "query:", query)
+        follow_up_values = []
+        for db_record in query_result:
+            if table_name not in selected_objects_dict.keys():
+                selected_objects_dict[table_name] = [db_record]
+            else:
+                if db_record not in selected_objects_dict[table_name]:
+                    selected_objects_dict[table_name].append(db_record)
+            if recursive:
+                db_object_columns = self.get_db_object_columns(db_record)
+                if recursive_key_column in db_object_columns:
+                    follow_up_values.append(db_record.__getattribute__(recursive_key_column))
+
+        new_query_values = "', '".join(follow_up_values)
+        if recursive and len(follow_up_values) > 0 and new_query_values != query_values:
+            self.get_db_objects(table_name, query_column, new_query_values, selected_objects_dict, recursive)
+        return selected_objects_dict
 
 
 class TE_Table_TreeWidgetItem(TemplateEditorTreeWidgetItem):
@@ -210,18 +283,56 @@ class TE_RelationColumn_TreeWidgetItem(TemplateEditorTreeWidgetItem):
     @property
     def display_name(self):
         display = None
+
         if isinstance(self.object_data, dict):
-            table_name = self.object_data.get("table_name", "")
-            column_name = self.object_data.get("column_name", "")
-            relation_type = self.object_data.get("relation_type", "")
-            follow_table = self.object_data.get("follow_table", "")
-            follow_column = self.object_data.get("follow_column", "")
-            display = f"{relation_type}: {column_name} -> {follow_table}"
+            if self.ParentTable == self.TableGroup:
+                display = f"{self.follow_table} - > {self.follow_column}"
+            else:
+                display = f"{self.follow_column} -> {self.follow_table}"
 
         if isinstance(self.object_data, Row):
             if self.objectkey_table == "DialogColumn":
-                display = f"{self.object_data.ColumnName} - ({self.object_data.Caption})"
+                display = f"{self.object_data.Caption} - ({self.object_data.ColumnName})"
         
         if display is None:
             display = "Column without display name"
+
         return display
+
+    @property
+    def follow_table(self):
+        if self.ParentTable == self.TableGroup:
+            return self.ChildTable
+        return self.ParentTable
+    
+    @property
+    def follow_column(self):
+        if self.ParentTable == self.TableGroup:
+            return self.ParentColumn
+        return self.ChildColumn
+
+    @property
+    def ChildTable(self):
+        if isinstance(self.object_data, dict):
+            return self.object_data.get("ChildTable", "")
+
+    @property
+    def ChildColumn(self):
+        if isinstance(self.object_data, dict):
+            return self.object_data.get("ChildColumn", "")
+
+    @property
+    def ParentTable(self):
+        if isinstance(self.object_data, dict):
+            return self.object_data.get("ParentTable", "")
+            
+    @property
+    def ParentColumn(self):
+        if isinstance(self.object_data, dict):
+            return self.object_data.get("ParentColumn", "")
+    
+    @property
+    def TableGroup(self):
+        if isinstance(self.object_data, dict):
+            return self.object_data.get("TableGroup", "")
+
